requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Configuration & Environment
pydantic-settings==2.1.0
python-dotenv==1.0.1
bcrypt==4.0.1

# Templates & Forms
jinja2==3.1.3
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1

# PDF Generation
xhtml2pdf==0.2.15
reportlab==4.0.9

# Storage (S3 support)
boto3==1.34.34

# Utilities
pytz==2024.1
